Add option to disable secret generation
This update introduces the ability to disable secret generation across the Aspirate Project. A "DisableSecrets" option is added, which when set to true, turns off all secret generations. This is particularly useful when secret generation is unnecessary or poses unnecessary overhead. The update also changes the Docker image for Redis to use the latest version, instead of a specific version.

+++ src/Aspirate.Commands/Actions/Manifests/ApplyManifestsToClusterAction.cs
@@ -66,6 +66,11 @@ public sealed class ApplyManifestsToClusterAction(
 
     private async Task WriteSecretsOutToTempFiles(List<string> files)
     {
+        if (CurrentState.DisableSecrets)
+        {
+            return;
+        }
+
         if (secretProvider is PasswordSecretProvider passwordSecretProvider)
         {
             Logger.MarkupLine($"\r\n[green]({EmojiLiterals.CheckMark}) Done:[/] Decrypting secrets from [blue]{CurrentState.InputPath}[/]");
@@ -92,6 +97,11 @@ public sealed class ApplyManifestsToClusterAction(
 
     private void CleanupSecretEnvFiles(IEnumerable<string> secretFiles)
     {
+        if (CurrentState.DisableSecrets)
+        {
+            return;
+        }
+
         foreach (var secretFile in secretFiles.Where(secretFile => fileSystem.File.Exists(secretFile)))
         {
             fileSystem.File.Delete(secretFile);
